fix(ui): make minimap workspace names readable
- rotate narrow-column workspace labels a quarter turn instead of
  stacking one character per line
- prefer the horizontal badge whenever the full name fits the strip
- stop the hover tooltip from collapsing into a glyph-wide strip
  (egui open-tooltip re-measure loop): unwrapped label + per-target id
- cap hover labels at 96 chars via a shared truncate_chars helper
- split label painting into app/minimap/labels.rs
